vertex-ai/iris_custom_training_artifact_registry/requirements.txt
pandas
scikit-learn
google-cloud-aiplatform
google-cloud-storage
fsspec
gcsfs